Split actions and fields into separate files
